+++ 0000_04_BlackJack21KKH/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>0000_04_BlackJack21KKH</name>
+	<name>0000_04_BlackJack21</name>
 	<comment></comment>
 	<projects>
 	</projects>
