+++ 0000_04_BlackJack21KKH/src/com/team21/blackjack/controller/BlackJackMain.java
@@ -1,0 +1,8 @@
+package com.team21.blackjack.controller;
+
+public class BlackJackMain {
+	public static void main(String[] args) {
+		Controller blackjack = new Controller();
+		blackjack.startGame();
+	}
+}
