+++ 0000_04_BlackJack21KKH/src/com/team21/blackjack/model/Player_Split.java
@@ -1,4 +1,4 @@
-package com.team21.hamoon;
+package com.team21.blackjack.model;
 
 import java.util.ArrayList;
 import java.util.List;
