+++ 0000_04_BlackJack21KKH/src/com/team21/blackjack/util/Scanner_Util.java
@@ -1,0 +1,42 @@
+package com.team21.blackjack.util;
+
+import java.util.InputMismatchException;
+import java.util.Scanner;
+
+public class Scanner_Util {
+
+	public int scanInt() {
+		int input = 0;
+		while (true) {
+			try {
+				Scanner scan = new Scanner(System.in);
+				input = scan.nextInt();
+				break;
+			} catch (InputMismatchException e) {
+				System.out.println("잘못된 입력입니다");
+			} 
+			
+		}
+		return input;
+	}
+	
+	public void scanAnything() {
+	
+		Scanner scan = new Scanner(System.in);
+		scan.nextLine();
+	}
+	//범위 안의 숫자만 입력받을 수 있는 메소드
+		public int input (int min, int max) {
+			int input;
+			while(true) {
+				input = scanInt();
+				if(input >= min && input <= max) {
+					break;
+				} 
+				System.out.println("범위 안의 숫자를 입력해주세요");
+			}
+			return input;
+		}
+		
+
+}
