--- 0000_04_BlackJack21KKH/src/com/team21/khkim/TestMain2.java
@@ -1,11 +1,0 @@
-package com.team21.khkim;
-
-
-public class TestMain2 {
-	public static void main(String[] args) {
-		Controller c = new Controller();
-		
-		
-		c.startGame();
-	}
-}
