--- 0000_04_BlackJack21KKH/src/com/team21/sy/Test_Main.java
@@ -1,18 +1,0 @@
-package com.team21.sy;
-
-import java.util.Scanner;
-
-public class Test_Main {
-
-	public static void main(String[] args) {
-
-		Money bb = new Money();
-		System.out.println("배팅 금액을 입력하세요.");
-//		bb.firstBet();	
-//		bb.doubleDown();
-		bb.Insurance();
-		System.out.println("수엽");
-		System.out.println("맛점");
-	}
-
-}
